replace useEffect with a separate RSC to grab data, replace stale appointments state with rows data from fetchAppointments, pass rows data to ReservationsTable, get rid of useStates and extraneous code
